+++ src/VirtoCommerce.ElasticAppSearch.Core/Models/Api/Search/Query/Facets/DateTimeRangeFacet.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace VirtoCommerce.ElasticAppSearch.Core.Models.Api.Search.Query.Facets
+{
+    public record DateTimeRangeFacet : RangeFacet<DateTime>
+    {
+    }
+}
