+++ src/VirtoCommerce.ElasticAppSearch.Core/Models/Api/Search/Query/Facets/Facet.cs
@@ -1,0 +1,16 @@
+﻿using Newtonsoft.Json;
+
+namespace VirtoCommerce.ElasticAppSearch.Core.Models.Api.Search.Query.Facets
+{
+    public record class Facet
+    {
+        [JsonRequired]
+        public virtual string Type { get; }
+
+        /// <summary>
+        /// Optional. Name given to facet.
+        /// </summary>
+        [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
+        public string Name { get; set; }
+    }
+}
