+++ src/VirtoCommerce.ElasticAppSearch.Core/Models/Api/Search/Query/Facets/FacetSort.cs
@@ -1,0 +1,11 @@
+﻿namespace VirtoCommerce.ElasticAppSearch.Core.Models.Api.Search.Query.Facets
+{
+    public record FacetSort : Field<SortOrder>
+    {
+        public FacetSort(FacetSortField field, SortOrder sortOrder)
+        {
+            FieldName = field.ToString().ToLowerInvariant();
+            Value = sortOrder;
+        }
+    }
+}
