+++ src/VirtoCommerce.ElasticAppSearch.Core/Models/Api/Search/Query/Facets/FacetSortField.cs
@@ -1,0 +1,8 @@
+﻿namespace VirtoCommerce.ElasticAppSearch.Core.Models.Api.Search.Query.Facets
+{
+    public enum FacetSortField
+    {
+        Count,
+        Value
+    }
+}
