+++ src/VirtoCommerce.ElasticAppSearch.Core/Models/Api/Search/Query/Facets/GeoLocationRangeFacet.cs
@@ -1,0 +1,13 @@
+﻿using Newtonsoft.Json;
+
+namespace VirtoCommerce.ElasticAppSearch.Core.Models.Api.Search.Query.Facets
+{
+    public record GeoLocationRangeFacet : RangeFacet<double>
+    {
+        [JsonRequired]
+        public GeoPoint Center { get; init; }
+
+        [JsonRequired]
+        public MeasurementUnit Unit { get; init; }
+    }
+}
