+++ src/VirtoCommerce.ElasticAppSearch.Core/Models/Api/Search/Query/Facets/NumberRangeFacet.cs
@@ -1,0 +1,6 @@
+﻿namespace VirtoCommerce.ElasticAppSearch.Core.Models.Api.Search.Query.Facets
+{
+    public record NumberRangeFacet : RangeFacet<double>
+    {
+    }
+}
