+++ src/VirtoCommerce.ElasticAppSearch.Core/Models/Api/Search/Result/FacetData.cs
@@ -1,0 +1,14 @@
+﻿namespace VirtoCommerce.ElasticAppSearch.Core.Models.Api.Search.Result;
+
+public record FacetData
+{
+    public string Name { get; init; }
+
+    public object Value { get; init; }
+
+    public object From { get; init; }
+
+    public object To { get; init; }
+
+    public int? Count { get; init; }
+}
