+++ src/VirtoCommerce.ElasticAppSearch.Core/Models/Api/Search/Result/SearchResult.cs
@@ -10,25 +10,3 @@ public record SearchResult
 
     public Dictionary<string, FacetResult[]> Facets { get; init; }
 }
-
-public record FacetResult
-{
-    public string Type { get; init; }
-
-    public string Name { get; init; }
-
-    public FacetData[] Data { get; init; }
-}
-
-public record FacetData
-{
-    public string Name { get; init; }
-
-    public object Value { get; init; }
-
-    public object From { get; init; }
-
-    public object To { get; init; }
-
-    public int? Count { get; init; }
-}
